Move `save_file` controller tests into FilesController tests.
We also removed and generate new cassettes
