--- src/api/spec/cassettes/Webui_PackageController/POST_save_file/as_ajax_request/adding_a_file_that_doesn_t_exist_yet/1_4_1_3_1.yml
@@ -1,303 +1,0 @@
----
-http_interactions:
-- request:
-    method: put
-    uri: http://backend:5352/source/home:tom/_meta?user=tom
-    body:
-      encoding: UTF-8
-      string: |
-        <project name="home:tom">
-          <title/>
-          <description/>
-          <person userid="tom" role="maintainer"/>
-        </project>
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - text/xml
-      Cache-Control:
-      - no-cache
-      Connection:
-      - close
-      Content-Length:
-      - '128'
-    body:
-      encoding: UTF-8
-      string: |
-        <project name="home:tom">
-          <title></title>
-          <description></description>
-          <person userid="tom" role="maintainer"/>
-        </project>
-    http_version: null
-  recorded_at: Fri, 29 May 2020 15:37:21 GMT
-- request:
-    method: put
-    uri: http://backend:5352/source/home:tom/my_package/_meta?user=tom
-    body:
-      encoding: UTF-8
-      string: |
-        <package name="my_package" project="home:tom">
-          <title>Dying of the Light</title>
-          <description>Illum eum nesciunt et.</description>
-        </package>
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - text/xml
-      Cache-Control:
-      - no-cache
-      Connection:
-      - close
-      Content-Length:
-      - '146'
-    body:
-      encoding: UTF-8
-      string: |
-        <package name="my_package" project="home:tom">
-          <title>Dying of the Light</title>
-          <description>Illum eum nesciunt et.</description>
-        </package>
-    http_version: null
-  recorded_at: Fri, 29 May 2020 15:37:21 GMT
-- request:
-    method: put
-    uri: http://backend:5352/source/home:tom/my_package/newly_created_file?user=tom
-    body:
-      encoding: UTF-8
-      string: some_content
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - text/xml
-      Cache-Control:
-      - no-cache
-      Connection:
-      - close
-      Content-Length:
-      - '203'
-    body:
-      encoding: UTF-8
-      string: |
-        <revision rev="2" vrev="2">
-          <srcmd5>5a8503c52a494e9a5660128cbdfa43f6</srcmd5>
-          <version>unknown</version>
-          <time>1590766641</time>
-          <user>tom</user>
-          <comment></comment>
-          <requestid/>
-        </revision>
-    http_version: null
-  recorded_at: Fri, 29 May 2020 15:37:21 GMT
-- request:
-    method: put
-    uri: http://backend:5352/source/home:tom/my_package/_meta?user=tom
-    body:
-      encoding: UTF-8
-      string: |
-        <package name="my_package" project="home:tom">
-          <title>Dying of the Light</title>
-          <description>Illum eum nesciunt et.</description>
-        </package>
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - text/xml
-      Cache-Control:
-      - no-cache
-      Connection:
-      - close
-      Content-Length:
-      - '146'
-    body:
-      encoding: UTF-8
-      string: |
-        <package name="my_package" project="home:tom">
-          <title>Dying of the Light</title>
-          <description>Illum eum nesciunt et.</description>
-        </package>
-    http_version: null
-  recorded_at: Fri, 29 May 2020 15:37:21 GMT
-- request:
-    method: get
-    uri: http://backend:5352/source/home:tom/my_package
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - text/xml
-      Cache-Control:
-      - no-cache
-      Connection:
-      - close
-      Content-Length:
-      - '207'
-    body:
-      encoding: UTF-8
-      string: |
-        <directory name="my_package" rev="2" vrev="2" srcmd5="5a8503c52a494e9a5660128cbdfa43f6">
-          <entry name="newly_created_file" md5="baab6c16d9143523b7865d46896e4596" size="12" mtime="1590766378"/>
-        </directory>
-    http_version: null
-  recorded_at: Fri, 29 May 2020 15:37:21 GMT
-- request:
-    method: get
-    uri: http://backend:5352/source/home:tom/my_package?view=info
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - text/xml
-      Cache-Control:
-      - no-cache
-      Connection:
-      - close
-      Content-Length:
-      - '228'
-    body:
-      encoding: UTF-8
-      string: |
-        <sourceinfo package="my_package" rev="2" vrev="2" srcmd5="5a8503c52a494e9a5660128cbdfa43f6" verifymd5="5a8503c52a494e9a5660128cbdfa43f6">
-          <error>bad build configuration, no build type defined or detected</error>
-        </sourceinfo>
-    http_version: null
-  recorded_at: Fri, 29 May 2020 15:37:21 GMT
-- request:
-    method: get
-    uri: http://backend:5352/source/home:tom/my_package
-    body:
-      encoding: US-ASCII
-      string: ''
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - text/xml
-      Cache-Control:
-      - no-cache
-      Connection:
-      - close
-      Content-Length:
-      - '207'
-    body:
-      encoding: UTF-8
-      string: |
-        <directory name="my_package" rev="2" vrev="2" srcmd5="5a8503c52a494e9a5660128cbdfa43f6">
-          <entry name="newly_created_file" md5="baab6c16d9143523b7865d46896e4596" size="12" mtime="1590766378"/>
-        </directory>
-    http_version: null
-  recorded_at: Fri, 29 May 2020 15:37:21 GMT
-- request:
-    method: post
-    uri: http://backend:5352/source/home:tom/my_package?cmd=diff&linkrev=base&onlyissues=1&orev=0&view=xml
-    body:
-      encoding: UTF-8
-      string: ''
-    headers:
-      Content-Type:
-      - application/octet-stream
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - text/xml
-      Content-Length:
-      - '298'
-      Cache-Control:
-      - no-cache
-      Connection:
-      - close
-    body:
-      encoding: UTF-8
-      string: |
-        <sourcediff key="7e48b6bf0da48f5476513ed993db0a02">
-          <old project="home:tom" package="my_package" rev="0" srcmd5="d41d8cd98f00b204e9800998ecf8427e"/>
-          <new project="home:tom" package="my_package" rev="2" srcmd5="5a8503c52a494e9a5660128cbdfa43f6"/>
-          <files/>
-          <issues>
-          </issues>
-        </sourcediff>
-    http_version: null
-  recorded_at: Fri, 29 May 2020 15:37:21 GMT
-recorded_with: VCR 5.1.0
