--- src/api/spec/cassettes/Webui_PackageController/POST_save_file/as_ajax_request/without_any_uploaded_file_data/fails_with_an_error_message.yml
@@ -1,83 +1,0 @@
----
-http_interactions:
-- request:
-    method: put
-    uri: http://backend:5352/source/home:tom/_meta?user=tom
-    body:
-      encoding: UTF-8
-      string: |
-        <project name="home:tom">
-          <title/>
-          <description/>
-          <person userid="tom" role="maintainer"/>
-        </project>
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - text/xml
-      Cache-Control:
-      - no-cache
-      Connection:
-      - close
-      Content-Length:
-      - '128'
-    body:
-      encoding: UTF-8
-      string: |
-        <project name="home:tom">
-          <title></title>
-          <description></description>
-          <person userid="tom" role="maintainer"/>
-        </project>
-    http_version: null
-  recorded_at: Fri, 29 May 2020 15:37:27 GMT
-- request:
-    method: put
-    uri: http://backend:5352/source/home:tom/my_package/_meta?user=tom
-    body:
-      encoding: UTF-8
-      string: |
-        <package name="my_package" project="home:tom">
-          <title>The Little Foxes</title>
-          <description>Ratione debitis pariatur blanditiis.</description>
-        </package>
-    headers:
-      Accept-Encoding:
-      - gzip;q=1.0,deflate;q=0.6,identity;q=0.3
-      Accept:
-      - "*/*"
-      User-Agent:
-      - Ruby
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Content-Type:
-      - text/xml
-      Cache-Control:
-      - no-cache
-      Connection:
-      - close
-      Content-Length:
-      - '158'
-    body:
-      encoding: UTF-8
-      string: |
-        <package name="my_package" project="home:tom">
-          <title>The Little Foxes</title>
-          <description>Ratione debitis pariatur blanditiis.</description>
-        </package>
-    http_version: null
-  recorded_at: Fri, 29 May 2020 15:37:27 GMT
-recorded_with: VCR 5.1.0
