+++ src/api/spec/controllers/webui/package_controller_spec.rb
@@ -148,115 +148,6 @@ RSpec.describe Webui::PackageController, vcr: true do
     end
   end
 
-  describe 'POST #save_file' do
-    RSpec.shared_examples 'tests for save_file action' do
-      before do
-        login(user)
-      end
-
-      context 'without any uploaded file data' do
-        it 'fails with an error message' do
-          do_request(project: source_project, package: source_package)
-          expect(response).to expected_failure_response
-          expect(flash[:error]).to eq("Error while creating '' file: No file or URI given.")
-        end
-      end
-
-      context 'with an invalid filename' do
-        it 'fails with a backend error message' do
-          do_request(project: source_project, package: source_package, filename: '.test')
-          expect(response).to expected_failure_response
-          expect(flash[:error]).to eq("Error while creating '.test' file: '.test' is not a valid filename.")
-        end
-      end
-
-      context "adding a file that doesn't exist yet" do
-        before do
-          do_request(project: source_project,
-                     package: source_package,
-                     filename: 'newly_created_file',
-                     file_type: 'local',
-                     file: 'some_content')
-        end
-
-        it { expect(response).to have_http_status(expected_success_status) }
-        it { expect(flash[:success]).to eq("The file 'newly_created_file' has been successfully saved.") }
-        it { expect(source_package.source_file('newly_created_file')).to eq('some_content') }
-      end
-
-      context 'uploading a utf-8 file' do
-        let(:file_to_upload) { File.read(File.expand_path(Rails.root.join('spec/support/files/chinese.txt'))) }
-
-        before do
-          do_request(project: source_project, package: source_package, filename: '学习总结', file: file_to_upload)
-        end
-
-        it { expect(response).to have_http_status(expected_success_status) }
-        it { expect(flash[:success]).to eq("The file '学习总结' has been successfully saved.") }
-
-        it 'creates the file' do
-          expect { source_package.source_file('学习总结') }.not_to raise_error
-          expect(URI.encode(source_package.source_file('学习总结'))).to eq(URI.encode(file_to_upload))
-        end
-      end
-
-      context 'uploading a file from remote URL' do
-        before do
-          do_request(project: source_project, package: source_package, filename: 'remote_file',
-                     file_url: 'https://raw.github.com/openSUSE/open-build-service/master/.gitignore')
-        end
-
-        after do
-          # Make sure the service only once get's created
-          source_package.destroy
-        end
-
-        it { expect(response).to have_http_status(expected_success_status) }
-        it { expect(flash[:success]).to eq("The file 'remote_file' has been successfully saved.") }
-
-        # Uploading a remote file creates a service instead of downloading it directly!
-        it 'creates a valid service file' do
-          expect { source_package.source_file('_service') }.not_to raise_error
-          expect { source_package.source_file('remote_file') }.to raise_error Backend::NotFoundError
-
-          created_service = source_package.source_file('_service')
-          expect(created_service).to eq(<<-EOT.strip_heredoc.strip)
-            <services>
-              <service name="download_url">
-                <param name="host">raw.github.com</param>
-                <param name="protocol">https</param>
-                <param name="path">/openSUSE/open-build-service/master/.gitignore</param>
-                <param name="filename">remote_file</param>
-              </service>
-            </services>
-          EOT
-        end
-      end
-    end
-
-    context 'as ajax request' do
-      let(:expected_success_status) { :ok }
-      let(:expected_failure_response) { have_http_status(:bad_request) }
-
-      def do_request(params)
-        post :save_file, xhr: true, params: params
-      end
-
-      include_examples 'tests for save_file action'
-    end
-
-    context 'as non-ajax request' do
-      let(:expected_success_status) { :found }
-      let(:expected_failure_response) { redirect_to(root_path) }
-
-      def do_request(params)
-        post :save_file, params: params
-      end
-
-      include_examples 'tests for save_file action'
-    end
-  end
-
   describe 'GET #show' do
     context 'require_package before_action' do
       context 'with an invalid package' do
